Implmented Create Booking Logic

+++ src/main/java/com/example/Uber/controller/LocationController.java
@@ -17,7 +17,7 @@ public class LocationController {
 
     @PostMapping("/driverLocation")
     public ResponseEntity<Boolean> saveDriverLocation(@RequestBody DriverLocationDTO driverLocation) {
-        Boolean saved = locationService.saveDriverLocation(driverLocation.getDriverId(), driverLocation.getLatitude(), driverLocation.getLongitude());
+        Boolean saved = locationService.saveDriverLocation(driverLocation.getDriverId().intValue(), driverLocation.getLatitude(), driverLocation.getLongitude());
         return ResponseEntity.ok(saved);
     }
